Add files via upload
f you find a bug or want to make a suggestion, then join my Discord server: https://discord.gg/pykzREcAuZ I will not consider adding features that are too cheaty, pointless or are just unrelated to this module.

IllegalMap is a ChatTriggers module that scans each Dungeon to give extra information to the user, like which rooms and puzzles are in the dungeon and exactly how many secrets there are before the dungeon even starts. The map also shows the user other useful information like the total secrets, death penalties, crypts, mimic and the current score right underneath the map so that the user never needs to look at the tab list. IllegalMap can be used on its own as your primary dungeon map, with significantly more features than any other map currently. IllegalMap is highly customizable through the /dmap command.

Note: Due to Hypixel's render distance limitation, not all of the dungeon will be visible from the starting room.